requirements.txt
python-telegram-bot
deep-translator
python-dotenv 